Refatorado codigo Game.js
