+++ src/services/storage.js
@@ -1,7 +1,5 @@
 export const getStorage = (key) => JSON.parse(localStorage.getItem(key));
 
-// export const setStorage = (key) =>
-
 export const setRankingStorage = (key, value) => {
   const keyStorage = getStorage(key);
   if (keyStorage) {
@@ -11,4 +9,11 @@ export const setRankingStorage = (key, value) => {
   } else {
     localStorage.setItem(key, JSON.stringify([value]));
   }
+};
+
+export const setScorePlayerStorage = (key, newScore) => {
+  const { player } = getStorage(key);
+  player.score += newScore;
+  player.assertions += 1;
+  localStorage.setItem('state', JSON.stringify({ player }));
 };
